requirements.txt
crawl4ai>=0.6.0
asyncpg>=0.29.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
elasticsearch==8.17.*
